reviewed and refactored some of the directories in mastermind
