+++ mastermind/Abstract/IRandomNumberGenerator.cs
@@ -1,4 +1,4 @@
-namespace mastermind.RandomNumberGenerator
+namespace mastermind.Abstract
 {
     public interface IRandomNumberGenerator
     {
