+++ mastermind/Hint/RandomHintIterator.cs
@@ -1,14 +1,14 @@
 using System.Collections.Generic;
-using mastermind.RandomNumberGenerator;
+using mastermind.Abstract;
 
-namespace mastermind.Iterators
+namespace mastermind.Hint
 {
-    public class RandomHintIterator : IIterator<Hint.Hint>
+    public class RandomHintIterator : IIterator<Hint>
     {
         private readonly IRandomNumberGenerator _generator;
-        private readonly List<Hint.Hint> _hints;
+        private readonly List<mastermind.Hint.Hint> _hints;
 
-        public RandomHintIterator(List<Hint.Hint> hints, IRandomNumberGenerator generator)
+        public RandomHintIterator(List<mastermind.Hint.Hint> hints, IRandomNumberGenerator generator)
         {
             _generator = generator;
             _hints = hints;
@@ -19,7 +19,7 @@ namespace mastermind.Iterators
             return _hints.Count != 0;
         }
 
-        public Hint.Hint GetNext()
+        public mastermind.Hint.Hint GetNext()
         {
             var randomIndex = _generator.NextRandom(_hints.Count);
             var randomHint = _hints[randomIndex];
