+++ mastermind/Program.cs
@@ -1,12 +1,14 @@
-﻿namespace mastermind
+﻿using mastermind.Game;
+
+namespace mastermind
 {
     internal static class Program
     {
         private static void Main()
         { 
-            var randomNumberGenerator = new RandomNumberGenerator.RandomNumberGenerator();
-            var console = new GameConsole.GameGameConsole();
-            var game = new Game.MastermindGame(console, randomNumberGenerator);
+            var randomNumberGenerator = new RandomNumberGenerator();
+            var console = new GameGameConsole();
+            var game = new Game.Game(console, randomNumberGenerator);
             game.Run();
         }
     }
