+++ mastermindTest/ColoursTest.cs
@@ -1,6 +1,5 @@
 using mastermind;
 using mastermind.Colours;
-using mastermind.RandomNumberGenerator;
 using Xunit;
 
 namespace mastermindTest
