+++ mastermindTest/GameDialogueTest.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using mastermind;
+using mastermind.Abstract;
 using mastermind.Colours;
 using mastermind.Game;
-using mastermind.GameConsole;
 using mastermind.Hint;
 using Moq;
 using Xunit;
