+++ mastermindTest/HintsTest.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using mastermind;
+using mastermind.Abstract;
 using mastermind.Colours;
 using mastermind.Hint;
-using mastermind.RandomNumberGenerator;
 using Moq;
 using Xunit;
